chore(agent): Move Agent interface to proper package
- moved SwabbieAgent to agent package in core
