+++ swabbie-core/src/main/kotlin/com/netflix/spinnaker/swabbie/AgentRunner.kt
@@ -18,7 +18,6 @@ package com.netflix.spinnaker.swabbie
 
 import com.netflix.spectator.api.Id
 import com.netflix.spectator.api.Registry
-import com.netflix.spinnaker.SwabbieAgent
 import com.netflix.spinnaker.swabbie.model.WorkConfiguration
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
